+++ packages/gamut/src/Form/SelectDropdown.tsx
@@ -1,6 +1,9 @@
-import { ArrowChevronDownIcon } from '@codecademy/gamut-icons';
-import { theme } from '@codecademy/gamut-styles';
-import { css } from '@emotion/react';
+import {
+  ArrowChevronDownIcon,
+  GamutIconProps,
+  MiniChevronDownIcon,
+} from '@codecademy/gamut-icons';
+import { useTheme } from '@emotion/react';
 import React, {
   ReactNode,
   SelectHTMLAttributes,
@@ -16,44 +19,85 @@ import ReactSelect, {
   StylesConfig,
 } from 'react-select';
 
-import { SelectComponentProps } from './Select';
+import { Box } from '../Box';
+import { SelectComponentProps, SelectOptions } from './Select';
 import {
-  colorStates,
-  conditionalBorderStyles,
-  formDropdownStyles,
-  formFieldStyles,
-} from './styles/shared';
-import { conditionalStyleProps } from './styles/shared-system-props';
+  conditionalBorderStates,
+  dropdownBorderStates,
+  dropdownBorderStyles,
+  optionBackground,
+  placeholderColor,
+  selectDropdownStyles,
+  sizeVariants,
+  textColor,
+} from './styles';
 import { parseOptions } from './utils';
 
-const { DropdownIndicator, SelectContainer } = SelectDropdownElements;
+export interface IconOption {
+  label: string;
+  value: string;
+  icon?: React.ComponentType<GamutIconProps>;
+}
+
+export interface SelectDropdownSizes {
+  size?: 'small' | 'medium';
+}
 
 type SelectDropdownBaseProps = Omit<
   SelectComponentProps,
-  'onChange' | 'defaultValue'
->;
+  'onChange' | 'defaultValue' | 'options'
+> &
+  SelectDropdownSizes;
+
+export type SelectDropdownOptions = SelectOptions | IconOption[];
+
 interface SelectDropdownProps
   extends SelectDropdownBaseProps,
-    Pick<NamedProps, 'onChange' | 'isSearchable'>,
+    Pick<NamedProps, 'onChange' | 'isSearchable' | 'onInputChange'>,
     Pick<SelectHTMLAttributes<HTMLSelectElement>, 'value' | 'disabled'> {
   inputProps?: Record<string, string | number | boolean>;
   name?: string;
   placeholder?: string;
+  options?: SelectDropdownOptions;
+  shownOptionsLimit?: 1 | 2 | 3 | 4 | 5 | 6;
 }
 
-type OptionStrict = {
+const { DropdownIndicator, SelectContainer } = SelectDropdownElements;
+
+export interface OptionStrict {
   label: string;
   value: string;
-};
+}
 
 type CustomContainerProps = ContainerProps<OptionStrict, false> & {
   children?: ReactNode[];
 };
 
-const ChevronDropdown = (props: IndicatorProps<OptionTypeBase, false>) => {
+const indicatorSizes = {
+  small: {
+    size: 12,
+    icon: MiniChevronDownIcon,
+  },
+  medium: {
+    size: 16,
+    icon: ArrowChevronDownIcon,
+  },
+};
+
+interface SizedIndicatorProps extends IndicatorProps<OptionTypeBase, false> {
+  selectProps: SelectDropdownSizes;
+}
+
+const ChevronDropdown = (props: SizedIndicatorProps) => {
+  const { size } = props.selectProps;
+  const color = props.isDisabled ? 'text-disabled' : 'text';
+  const { icon: IndicatorIcon, ...iconProps } = indicatorSizes[
+    size ?? 'medium'
+  ];
+
   return (
     <DropdownIndicator {...props}>
-      <ArrowChevronDownIcon size={16} />
+      <IndicatorIcon {...iconProps} color={color} />
     </DropdownIndicator>
   );
 };
@@ -70,85 +114,20 @@ const CustomContainer = ({ children, ...rest }: CustomContainerProps) => {
   );
 };
 
-const selectBaseStyles = ({
-  error,
-  activated,
-  isFocused,
-  isDisabled,
-}: conditionalStyleProps) => css`
-  ${formFieldStyles}
-  ${conditionalBorderStyles({ error, activated, isFocused, isDisabled })}
-  line-height: ${theme.lineHeight.base};
-  display: flex;
-`;
-
-const errorColorState = (error: boolean) => {
-  const color = error ? colorStates.error.color : colorStates.base.color;
-  return color;
-};
-
-const customStyles: StylesConfig<OptionTypeBase, false> = {
-  container: (provided, state) => ({
-    ...provided,
-    pointerEvents: 'visible',
-    cursor: state.selectProps.isSearchable ? 'text' : 'pointer',
-    width: '100%',
-    minWidth: '7rem',
-  }),
-
-  control: (provided, state) => ({
-    ...selectBaseStyles({
-      error: state.selectProps.error,
-      activated: state.selectProps.activated,
-      isFocused: state.isFocused,
-      isDisabled: state.isDisabled,
-    }),
-  }),
-
-  dropdownIndicator: (provided, state) => ({
-    color: errorColorState(state.selectProps.error),
-    display: 'flex',
-    padding: '0',
-    pointerEvents: 'none',
-  }),
-
-  input: (provided, state) => ({
-    padding: '0',
-    margin: '0',
-  }),
-
-  menu: (provided, state) => ({
-    ...provided,
-    ...formDropdownStyles(state.selectProps.error),
-  }),
-
-  option: (provided, state) => ({
-    padding: '14px 11px 14px 11px',
-    cursor: 'pointer',
-    backgroundColor:
-      (state.isSelected && colorStates.dropdown.selected.backgroundColor) ||
-      (state.isFocused && colorStates.dropdown.focused.backgroundColor) ||
-      'transparent',
-    '&:hover': {
-      backgroundColor: colorStates.dropdown.focused.backgroundColor,
-    },
-  }),
-
-  singleValue: (provided, state) => ({
-    color: errorColorState(state.selectProps.error),
-    display: 'flex',
-  }),
-
-  valueContainer: (provided) => ({
-    ...provided,
-    padding: 0,
-  }),
+const formatOptionLabel = ({ label, icon: Icon, size }: any) => {
+  return (
+    <Box display="flex" alignItems="center">
+      {Icon && <Icon size={size === 'small' ? 16 : 24} color="text" ml={4} />}
+      <Box as="span" pl={Icon ? 16 : 0}>
+        {label}
+      </Box>
+    </Box>
+  );
 };
 
 const defaultProps = {
   name: undefined,
   isMulti: false,
-  styles: customStyles,
   components: {
     DropdownIndicator: ChevronDropdown,
     IndicatorSeparator: () => null,
@@ -167,6 +146,8 @@ export const SelectDropdown: React.FC<SelectDropdownProps> = ({
   placeholder = 'Select an option',
   inputProps,
   isSearchable = false,
+  shownOptionsLimit = 6,
+  size,
   ...rest
 }) => {
   const [activated, setActivated] = useState(false);
@@ -180,9 +161,88 @@ export const SelectDropdown: React.FC<SelectDropdownProps> = ({
     setActivated(true);
   };
 
+  const theme = useTheme();
+
+  const memoizedStyles: StylesConfig<OptionTypeBase, false> = useMemo(() => {
+    return {
+      container: (provided, state) => ({
+        ...provided,
+        pointerEvents: 'visible',
+        cursor: state.selectProps.isSearchable ? 'text' : 'pointer',
+        width: '100%',
+        minWidth: '7rem',
+      }),
+
+      control: (provided, state) => {
+        return {
+          ...selectDropdownStyles({ theme }),
+          ...sizeVariants({ size: state.selectProps.size, theme }),
+          ...conditionalBorderStates({
+            isFocused: state.isFocused,
+            isDisabled: state.isDisabled,
+            error: state.selectProps.error,
+            activated: state.selectProps.activated,
+            theme,
+          }),
+        };
+      },
+
+      dropdownIndicator: (provided, state) => ({
+        color: 'currentColor',
+        display: 'flex',
+        padding: '0',
+        pointerEvents: 'none',
+      }),
+
+      input: (provided, state) => ({
+        ...textColor({ theme }),
+        padding: '0',
+        margin: '0',
+      }),
+
+      menu: (provided, state) => ({
+        ...provided,
+        ...dropdownBorderStyles({ theme }),
+        ...dropdownBorderStates({ error: state.selectProps.error, theme }),
+      }),
+
+      menuList: (provided, state) => {
+        const sizeInteger = state.selectProps.size === 'small' ? 2 : 3;
+        const maxHeight = `${
+          state.selectProps.shownOptionsLimit * sizeInteger
+        }rem`;
+        return {
+          ...provided,
+          maxHeight,
+        };
+      },
+
+      placeholder: (provided, state) => ({
+        ...provided,
+        ...placeholderColor({ theme }),
+      }),
+
+      option: (provided, state) => ({
+        padding: state.selectProps.size === 'small' ? '3px 14px' : '11px 14px',
+        cursor: 'pointer',
+        ...optionBackground(state.isSelected, state.isFocused)({ theme }),
+      }),
+
+      singleValue: (provided, state) => ({
+        ...textColor({ theme }),
+        display: 'flex',
+      }),
+
+      valueContainer: (provided) => ({
+        ...provided,
+        padding: 0,
+      }),
+    };
+  }, [theme]);
+
   const selectOptions = useMemo(() => {
-    return parseOptions({ options, id });
-  }, [options, id]);
+    return parseOptions({ options, id, size });
+  }, [options, id, size]);
 
   const parsedValue = useMemo(() => {
     const currentValue = selectOptions.find(
@@ -199,12 +259,16 @@ export const SelectDropdown: React.FC<SelectDropdownProps> = ({
       value={parsedValue}
       activated={activated}
       error={Boolean(error)}
+      formatOptionLabel={formatOptionLabel}
       onChange={changeHandler}
       inputProps={{ ...inputProps, ...baseInputProps }}
       isDisabled={disabled}
       options={selectOptions}
       placeholder={placeholder}
+      styles={memoizedStyles}
       isSearchable={isSearchable}
+      size={size}
+      shownOptionsLimit={shownOptionsLimit}
       {...rest}
     />
   );