+++ packages/gamut/src/GridForm/GridFormButtons/index.tsx
@@ -1,22 +1,24 @@
+import { ColorModes } from '@codecademy/gamut-styles';
 import React, { ComponentProps } from 'react';
 import { Mode, useFormContext } from 'react-hook-form';
 
 import { Box } from '../../Box';
 import { CTAButton, FillButton, TextButton } from '../../Button';
 import { ButtonProps } from '../../Button/shared';
+import { SubmitButton, SubmitButtonProps } from '../../Form';
 import { Column } from '../../Layout';
-import { VisualTheme } from '../../theming/VisualTheme';
 
 export type GridFormButtonsPosition = keyof typeof positions;
 
-export type SubmitButtonType = 'cta' | 'fill';
+export type SubmitButtonType = keyof typeof buttonMap;
 
 export type GridFormSubmitProps = {
   contents: React.ReactNode;
   position?: GridFormButtonsPosition;
   size: ComponentProps<typeof Column>['size'];
-  disabled?: ButtonProps['disabled'];
-  mode?: VisualTheme;
+  disabled?: SubmitButtonProps['disabled'];
+  loading?: SubmitButtonProps['loading'];
+  mode?: ColorModes;
   type?: SubmitButtonType;
   validation?: Mode;
 };
@@ -38,40 +40,14 @@ const positions = {
   stretch: 'stretch',
 };
 
+const buttonMap = {
+  cta: CTAButton,
+  fill: FillButton,
+};
+
 export const GridFormButtons: React.FC<
   GridFormSubmitProps & CancelButtonProps
-> = ({
-  validation,
-  type,
-  mode,
-  disabled,
-  contents,
-  size,
-  position,
-  cancel,
-}) => {
-  const { formState } = useFormContext();
-  const isDisabled =
-    (validation === 'onChange' && !formState.isValid) || disabled;
-
-  const getButton = () => {
-    switch (type) {
-      case 'cta':
-        return (
-          <CTAButton type="submit" mode={mode} disabled={isDisabled}>
-            {contents}
-          </CTAButton>
-        );
-
-      default:
-        return (
-          <FillButton type="submit" mode={mode} disabled={isDisabled}>
-            {contents}
-          </FillButton>
-        );
-    }
-  };
-
+> = ({ type = 'fill', ...props }) => {
   return (
     <Column size={size}>
       <Box
@@ -82,7 +58,14 @@ export const GridFormButtons: React.FC<
         {cancel && (
           <TextButton {...cancel} mr={32} data-testid="cancel-button" />
         )}
-        {getButton()}
+        <SubmitButton
+          as={buttonMap[type]}
+          mode={props.mode}
+          disabled={props.disabled}
+          loading={props.loading}
+        >
+          {props.contents}
+        </SubmitButton>
       </Box>
     </Column>
   );