+++ packages/styleguide/stories/Atoms/FormInputs/examples.tsx
@@ -6,9 +6,18 @@ import {
   Column,
   FormGroup,
   LayoutGrid,
+  SelectDropdown,
 } from '@codecademy/gamut/src';
+import { RadarIcon, ResponsiveIcon, RocketIcon } from '@codecademy/gamut-icons';
+import {
+  LockIcon,
+  NetworkUserIcon,
+  PeopleIcon,
+} from '@codecademy/gamut-icons/src';
 import { Background, ColorMode } from '@codecademy/gamut-styles/src';
 import React, { ChangeEvent, useState } from 'react';
+
+import { ColorModeExampleWrapper } from '../../Foundations/ColorMode/examples';
 
 type CustomCheckboxProps = Omit<CheckboxProps, 'checked'> & {
   defaultChecked?: boolean;
@@ -43,7 +52,7 @@ const CustomCheckbox: React.FC<CustomCheckboxProps> = ({
   );
 };
 
-export const CheckboxExample = () => {
+export const CheckboxExample: React.FC = () => {
   return (
     <ColorMode mode="light">
       <Box bg="background" border={1} p={16}>
@@ -74,5 +83,126 @@ export const CheckboxExample = () => {
         </LayoutGrid>
       </Box>
     </ColorMode>
+  );
+};
+
+export const SelectDropdownIconExample: React.FC = () => (
+  <Box p={16} width="100%" height="20rem" border={1} bg="palePink">
+    <FormGroup label="i have pizzazz">
+      <SelectDropdown
+        options={[
+          {
+            label: 'ohai',
+            value: 'ohai',
+            icon: RocketIcon,
+          },
+          {
+            label: 'surprise pacman',
+            value: 'pacman',
+            icon: RadarIcon,
+          },
+          {
+            label: `who's that pokemon?`,
+            value: 'what',
+            icon: ResponsiveIcon,
+          },
+        ]}
+      />
+    </FormGroup>
+    <Background p={16} bg="navy">
+      <FormGroup label="i am fancy">
+        <SelectDropdown
+          size="small"
+          options={[
+            {
+              label: 'ohai',
+              value: 'ohai',
+              icon: RocketIcon,
+            },
+            {
+              label: 'surprise pacman',
+              value: 'pacman',
+              icon: RadarIcon,
+            },
+            {
+              label: `who's that pokemon?`,
+              value: 'what',
+              icon: ResponsiveIcon,
+            },
+          ]}
+        />
+      </FormGroup>
+    </Background>
+  </Box>
+);
+
+export const SelectDropdownExample: React.FC = () => {
+  return (
+    <ColorModeExampleWrapper>
+      <Box height="34rem">
+        <FormGroup label="i have a fancy colormode" htmlFor="colormode">
+          <SelectDropdown
+            aria-label="colormode"
+            options={['ooo fancy', 'yes I am!', ':)']}
+            name="colormode"
+            isSearchable
+          />
+        </FormGroup>
+        <FormGroup
+          label="i have an error"
+          htmlFor="colormode-error"
+          error="cry cry cry"
+        >
+          <SelectDropdown
+            aria-label="colormode-error"
+            options={['ooo fancy', 'yes I am!', ':)']}
+            name="colormode-error"
+            isSearchable
+            error
+          />
+        </FormGroup>
+        <FormGroup label="i am disabled" htmlFor="colormode-disabled">
+          <SelectDropdown
+            aria-label="colormode-disabled"
+            options={['ooo fancy', 'yes I am!', ':)']}
+            name="colormode-disabled"
+            isSearchable
+            disabled
+          />
+        </FormGroup>
+        <FormGroup label="i have icons" htmlFor="colormode-icons">
+          <SelectDropdown
+            aria-label="icon"
+            options={[
+              {
+                label: 'Private: Only you can prevent forest fires',
+                value: 'private',
+                icon: LockIcon,
+              },
+              {
+                label: 'Public: Wait what?',
+                value: 'public',
+                icon: NetworkUserIcon,
+              },
+              {
+                label: 'Monsters: Monsters are such interesting people',
+                value: 'monsters',
+                icon: PeopleIcon,
+              },
+            ]}
+            name="colormode-icons"
+          />
+        </FormGroup>
+        <FormGroup label="i am smol" htmlFor="colormode-small">
+          <SelectDropdown
+            aria-label="colormode-small"
+            options={['small', 'yes', 'very very']}
+            size="small"
+            name="colormode-small"
+            isSearchable
+          />
+        </FormGroup>
+      </Box>
+    </ColorModeExampleWrapper>
   );
 };